RECNO: widen sparsemap index to 64-bit and namespace symbols
The RECNO dirty-page map keys each entry by ((relid << 32) | blockno),
a composite value that routinely exceeds 2^32.  The bitmap's chunk-start
index type and the cursor's cached start-index field were uint32, so any
key at or above the 4 GB boundary was silently truncated.  The cursor
then matched the wrong chunk, set bits vanished, and parallel crash
recovery aborted before reaching a consistent state.

Widen __sm_idx_t and sparsemap.m_cursor_start_idx to 64-bit so the full
composite key space is representable.  The per-chunk overhead and the
empty-map size grow from 4 to 8 bytes accordingly; the lifecycle test
oracles are updated to expect sizeof(uint64).  The sm_scan callback
delivers absolute 64-bit bit positions, so its buffer and the
sparsemap_scan wrapper are uint64; the test callback follows.

Adopt the upstream SPARSEMAP_PREFIX mechanism with prefix "pg_" so every
exported type, struct tag, and function is namespaced (pg_sm_create,
struct pg_sparsemap, ...) and cannot collide with other sm_* code linked
into the backend.  Track upstream sparsemap v4.0.0: this is the release
that widens the chunk-start offset to 64 bits as its headline fix, and
also adds the cast and unsigned-literal hardening carried here.

+++ src/test/modules/test_sparsemap/test_sparsemap.c
@@ -85,14 +85,15 @@ test_lifecycle(void)
 	map = sparsemap_create(1024);
 	EXPECT_TRUE(map != NULL);
 	EXPECT_EQ_SZ(sparsemap_get_capacity(map), 1024);
-	EXPECT_EQ_SZ(sparsemap_get_size(map), sizeof(uint32));
+	/* empty-map size equals the per-chunk offset width (__sm_idx_t == uint64) */
+	EXPECT_EQ_SZ(sparsemap_get_size(map), sizeof(uint64));
 	sparsemap_free(map);
 
 	/* init with caller-provided buffer */
 	buf = palloc0(1024);
 	sparsemap_init(&local, buf, 1024);
 	EXPECT_EQ_SZ(sparsemap_get_capacity(&local), 1024);
-	EXPECT_EQ_SZ(sparsemap_get_size(&local), sizeof(uint32));
+	EXPECT_EQ_SZ(sparsemap_get_size(&local), sizeof(uint64));
 
 	/* add some data, then clear */
 	sparsemap_add(&local, 42);
@@ -379,7 +380,7 @@ static size_t scan_count;
 static uint64 scan_last_idx;
 
 static void
-scan_counter(uint32 v[], size_t n, void *aux)
+scan_counter(uint64 v[], size_t n, void *aux)
 {
 	(void) aux;
 	for (size_t i = 0; i < n; i++)
